partials: allow map partial re-use and setting zoom level
The maps partial might be useful not only on the contact page, but also
on individual pages with per-page parameters. This modification combined
with the previous custom partials enables this.

Now, a latitude/longitude setting can be set per page. An example could
be:
---
latitude: "0"
longitude: "0"
direction: 1,1
map_zoom: 17
custom_partials:
  end: map.html
---

This will embed a map at the bottom of the page, just like in the
contact page, and use the per-page instead of the global config
settings.