extend navigation highlighting to taxonomy list pages (#220)
This adds the same kind of navigation highlighting as the PRs #191 and #211
when on [taxonomy list pages](https://gohugo.io/content-management/taxonomies/#default-destinations)
(e.g. `/categories/programming/` or `/tags/hugo/`), i.e. the selected tag or
category will be highlighted in the color of the chosen style (default, blue,
green, etc.)